% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-decode.R
\name{.bandpower_transform}
\alias{.bandpower_transform}
\title{Band-power (log-variance of band-filtered signal) features - fit-free}
\usage{
.bandpower_transform(data, idx, sr, bands)
}
\description{
Band-power (log-variance of band-filtered signal) features - fit-free
}
\keyword{internal}
