% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-timefreq.R
\name{.compute_dpss}
\alias{.compute_dpss}
\title{Compute DPSS (Slepian) tapers via tridiagonal eigendecomposition}
\usage{
.compute_dpss(n, nw, k)
}
\arguments{
\item{n}{Integer length of the data.}

\item{nw}{Numeric time-half-bandwidth product.}

\item{k}{Integer number of tapers to return.}
}
\value{
Matrix of dimension n x k containing the DPSS tapers as columns.
}
\description{
Approximates Discrete Prolate Spheroidal Sequences by constructing a
symmetric tridiagonal matrix whose eigenvectors corresponding to the
largest eigenvalues are the desired tapers.
}
\keyword{internal}
