% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-decode.R
\name{.csp_fit}
\alias{.csp_fit}
\title{Fit CSP spatial filters on a set of training trials}
\usage{
.csp_fit(data, idx, labels, n_filters = 3)
}
\arguments{
\item{data}{3D array (time x channels x trials).}

\item{idx}{Training trial indices.}

\item{labels}{Class labels for \code{idx} (length == \code{length(idx)}).}

\item{n_filters}{Number of CSP filter pairs (default: 3).}
}
\value{
A \code{(2 n_filters) x channels} spatial filter matrix.
}
\description{
Fit CSP spatial filters on a set of training trials
}
\keyword{internal}
