% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-decode.R
\name{.csp_transform}
\alias{.csp_transform}
\title{CSP log-variance features for a set of trials}
\usage{
.csp_transform(data, idx, W)
}
\description{
CSP log-variance features for a set of trials
}
\keyword{internal}
