% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-ica.R
\name{.fastica}
\alias{.fastica}
\title{FastICA algorithm (internal)}
\usage{
.fastica(whitened, n_components, max_iter, tol)
}
\arguments{
\item{whitened}{Whitened data matrix (n_time x n_components).}

\item{n_components}{Number of components.}

\item{max_iter}{Maximum iterations.}

\item{tol}{Convergence tolerance.}
}
\value{
Unmixing matrix W (n_components x n_components).
}
\description{
FastICA algorithm (internal)
}
\keyword{internal}
