% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-connectivity.R
\name{.hanning_window}
\alias{.hanning_window}
\title{Compute Hanning window}
\usage{
.hanning_window(n)
}
\arguments{
\item{n}{Window length.}
}
\value{
Numeric vector of Hanning window coefficients.
}
\description{
Compute Hanning window
}
\keyword{internal}
