% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-ica.R
\name{.jade}
\alias{.jade}
\title{JADE ICA algorithm (internal)}
\usage{
.jade(whitened, n_components, max_iter, tol)
}
\arguments{
\item{whitened}{Whitened data matrix (n_time x n_components).}

\item{n_components}{Number of components.}

\item{max_iter}{Maximum iterations.}

\item{tol}{Convergence tolerance.}
}
\value{
Unmixing matrix W (n_components x n_components).
}
\description{
Joint Approximate Diagonalization of Eigenmatrices. Computes fourth-order
cumulant matrices and performs joint approximate diagonalization using
Jacobi rotations.
}
\keyword{internal}
