% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-microstate.R
\name{.microstate_aahc}
\alias{.microstate_aahc}
\title{AAHC (Atomize and Agglomerate Hierarchical Clustering)}
\usage{
.microstate_aahc(topos, n_states)
}
\description{
AAHC (Atomize and Agglomerate Hierarchical Clustering)
}
\keyword{internal}
