% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-microstate.R
\name{.microstate_backfit}
\alias{.microstate_backfit}
\title{Backfit topographies to microstate maps}
\usage{
.microstate_backfit(data, maps)
}
\description{
Backfit topographies to microstate maps
}
\keyword{internal}
