% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-microstate.R
\name{.microstate_kmeans}
\alias{.microstate_kmeans}
\title{Polarity-invariant K-means for microstate clustering}
\usage{
.microstate_kmeans(topos, n_states, max_iter = 100)
}
\description{
Polarity-invariant K-means for microstate clustering
}
\keyword{internal}
