% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-timefreq.R
\name{.morlet_power_2d}
\alias{.morlet_power_2d}
\title{Compute Morlet wavelet power for 2D data (internal)}
\usage{
.morlet_power_2d(data, sr, frequencies, n_cycles)
}
\arguments{
\item{data}{Numeric matrix (time x channels).}

\item{sr}{Sampling rate in Hz.}

\item{frequencies}{Numeric vector of frequencies.}

\item{n_cycles}{Number of wavelet cycles.}
}
\value{
3D array (time x frequencies x channels) of power values.
}
\description{
Computes wavelet power (|W|^2) for a 2D matrix (time x channels) at
specified frequencies using Morlet wavelets with FFT convolution.
}
\keyword{internal}
