% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-vis.R
\name{.nose_shape}
\alias{.nose_shape}
\title{Generate nose shape coordinates}
\usage{
.nose_shape()
}
\value{
A data.frame with columns x, y.
}
\description{
Returns coordinates for a triangle at the top of the head circle
representing the nose.
}
\keyword{internal}
