% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-decode.R
\name{.riemann_fit}
\alias{.riemann_fit}
\title{Fit the Riemannian tangent-space reference (train Frechet mean)}
\usage{
.riemann_fit(data, idx)
}
\description{
Fit the Riemannian tangent-space reference (train Frechet mean)
}
\keyword{internal}
