% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeg-decode.R
\name{.riemann_transform}
\alias{.riemann_transform}
\title{Riemannian tangent-space features w.r.t. a fixed reference}
\usage{
.riemann_transform(data, idx, C_ref)
}
\description{
Riemannian tangent-space features w.r.t. a fixed reference
}
\keyword{internal}
